# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mentora",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://mentora-me.herokuapp.com",
      "variables": {}
    },
    {
      "url": "http://localhost:3001",
      "variables": {}
    }
  ],
  "paths": {
...
    "/mentor": {
      "get": {
        "tags": [
          "Mentor"
        ],
        "summary": "ListMentor",
        "operationId": "ListMentor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "https://mentora-me.herokuapp.com",
            "variables": {}
          }
        ]
      },
      "post": {
        "tags": [
          "Mentor"
        ],
        "summary": "CreateMentor",
        "operationId": "CreateMentor",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateMentorRequest"
                  },
                  {
                    "example": {
                      "name": "Joao",
                      "chavePix": "24535112333",
                      "phone": "+5562889646868",
                      "publicEmail": "joao@gmail.com",
                      "cnpj": "12111668910",
                      "mentoringArea": "PROGRAMAÇÃO"
                    }
                  }
                ]
              },
              "example": {
                "name": "Joao",
                "chavePix": "24535112333",
                "phone": "+5562889646868",
                "publicEmail": "joao@gmail.com",
                "cnpj": "12111668910",
                "mentoringArea": "PROGRAMAÇÃO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "https://mentora-me.herokuapp.com",
            "variables": {}
          }
        ]
      },
      "put": {
        "tags": [
          "Mentor"
        ],
        "summary": "UpdateMentor",
        "operationId": "UpdateMentor",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateMentorRequest"
                  },
                  {
                    "example": {
                      "name": "Joao das Neves",
                      "chavePix": "24535112333",
                      "phone": "+5562889646868",
                      "publicEmail": "joao@gmail.com",
                      "cnpj": "12111668910",
                      "mentoringArea": "PROGRAMAÇÃO"
                    }
                  }
                ]
              },
              "example": {
                "name": "Joao das Neves",
                "chavePix": "24535112333",
                "phone": "+5562889646868",
                "publicEmail": "joao@gmail.com",
                "cnpj": "12111668910",
                "mentoringArea": "PROGRAMAÇÃO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "https://mentora-me.herokuapp.com",
            "variables": {}
          }
        ]
      },
      "delete": {
        "tags": [
          "Mentor"
        ],
        "summary": "DeleteMentor",
        "operationId": "DeleteMentor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "https://mentora-me.herokuapp.com",
            "variables": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMentorRequest": {
        "title": "CreateMentorRequest",
        "required": [
          "name",
          "chavePix",
          "phone",
          "publicEmail",
          "cnpj",
          "mentoringArea"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "chavePix": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "publicEmail": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "mentoringArea": {
            "type": "string"
          }
        },
        "example": {
          "name": "Joao",
          "chavePix": "24535112333",
          "phone": "+5562889646868",
          "publicEmail": "joao@gmail.com",
          "cnpj": "12111668910",
          "mentoringArea": "PROGRAMAÇÃO"
        }
      },
      "UpdateMentorRequest": {
        "title": "UpdateMentorRequest",
        "required": [
          "name",
          "chavePix",
          "phone",
          "publicEmail",
          "cnpj",
          "mentoringArea"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "chavePix": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "publicEmail": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "mentoringArea": {
            "type": "string"
          }
        },
        "example": {
          "name": "Joao das Neves",
          "chavePix": "24535112333",
          "phone": "+5562889646868",
          "publicEmail": "joao@gmail.com",
          "cnpj": "12111668910",
          "mentoringArea": "PROGRAMAÇÃO"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}